fix: Replaced spaces with underscores in g_ExtractGLMVolumes saveoption
When calling commands through connector, spaces within parameter values
are not processed correctly. Also, by convention, individual parameter
/ option values (i.e. not lists or arbitrary strings) should not have
spaces. For this reason spaces in `by effect`, `by session`, and
`effect files` options for the saveoption parameter were replaced with
underscores in documentation.

In code itself the change was implemented by replacing underscores with
spaces (line 86), so that new parameter values with underscores are
converted to old expected values and used correctly in the rest of the
code that remains unchanged. In this way saveoption values can still be
used in without change when printing verbose reports (e.g. lines 179,
186). Additionally this also provides backward compatibility for any
code that passes the options with spaces instead of underscores.

Additional issues were identified and fixed in testing:

1/ If effects parameter was specified using an empty string (''), no
effects were extracted. The code now checks whether the effects
parameter is an empty string (line 102) so that the empty parameter
is used correctly when calling img_ExtractGLMEstimates and all effects
are exported.

2/ Code incorrectly checked for `by estimate` saveoption when it should
check for `by effect`, and the appropriate sorting was skipped
(line 156).

3/ A check for correct saveoption parameter was added. The function now
warns if an invalid option is used and exits (line 87).

trello-card: https://trello.com/c/oEGstAhK