feat: specify which sequence specific information to use
During `importHCP` and `importDICOM` a number of sequence specific
information such as EchoSpacing, DwellTime, PEDirection can be
extracted and added to the `sessions.txt` file that are later also
available in the batch file. If for any reason these information are
not valid, they should not be used.

This commit adds the ability to explicitly specify which sequence
specific inline information to use. It adds a new parameter
`use_sequence_info` that is set to 'all' by default, it adds a function
`checkInlineParameterUse` that simplifies checking whether a specific
information should be used for a specific image modality, and
implements the check for the use of inline sequence specific
information when setting up parameters to run PreFreeSurfer, fMRIVolume
and Diffusion HCP pipelines.

trello-card: https://trello.com/c/EHHIIKYU