fix: ensure import of SE-FM and FM images from HCP1200 dataset
importHCP failed to import SpinEchoFieldMap and FieldMap images from
HCP1200 dataset. The specification in importHCP.txt and code was adjusted
to correctly identify LR and RL SpinEchoFieldMap files as well as
FieldMap-Magnitude and FieldMap-Phase images. For the latter the code also
correctly assigns the fm([N]) to both the FieldMap pair of images as well
as the structural or functional images the FieldMap images pertain to.

The code still relies on unique names for SpinEchoFieldMap and FieldMap
image pairs. If different pairs use the same name, only one of the images
will be mapped and used.