fix: check whether a folder needs to be created
When a listfile was specified, the code allways checked for existence
and attempted to create the directory where the listfile was to be
saved, even though the path was empty as the listfile was supposed
to be created in the current folder.

The code now checks if a path was specified before checking for its
existence and trying to create it.

trello-card: https://trello.com/c/i7IOx4fH