feat: add the ability for processing different bold files and variates
The processing workflow code assumed that just one correct version of the
volume representation of the BOLD image existed, named using the formula
<boldname>[N].nii.gz. BOLD image statistics and dvars computation, as well
as nuisance signal extration were all based on this image. And they were
used for both the volume image as well as the cifti representation of the
same data.

It is though possible that a volume image is further processed in
HCPpipelines (and in future in other processing pipelines), resulting in
additional versions of the volume representation. For instance, after
running ICAFIX, *_hp2000_clean.nii.gz versions of BOLD files are generated
and based on them *_Atlas_hp2000_clean.nii.gz cifti files. It is important
that for these volume images and their cifti representations the correct
statistics are computed, the correct nuisance signals are extracted and
even the correct bold brain masks are created.

For this reason it is now possible to specify not only the desired cifti
file tail to be used when working with cifti data, but also the relevant
nifti (volume) file tail to be used both for volume and cifti data. This
tail is taken into account when computing and using bold statistics,
computing movement scrubbing, extracting nuisance data, creating bold brain
masks.

In addition the code now correctly maps the data from the hcp folder and
uses it in later processing, enabling general image and functional parallel
workflows. More specifically, parallel variants of hcp processing,
identified by hcp_suffix:

<session id>/hcp/<session id><hcp_suffix>

e.g.:

OP270/hcp/OP270_sulc

can now be mapped to and used in separate image folders specified using
img_suffix parameter:

<session id>/images<img_suffix>

e.g.

OP270/images_sulc

In addition, it is possible to map from and use files from different
variant of the MNINonLinear/Results folder using the hcp_bold_variant
and bold_variant parameters, which maps the functional data from:

<session id>/hcp/<session id><hcp_suffix>/MNINonLinear/Results<hcp_bold_variant>

to:

<session id>/images<img_suffix>/functional<bold_variant>

This enables separable parallel procesing and analysis workflows of
bold data.

trello-card: https://trello.com/c/CBVBi3WU