feat: add QuNex side [qx_]citi_tail and [qx_]nifti_tail parameters
To specify the volume and cifti representation of the bold data to work
with, QuNex used `hcp_cifti_tail` and `hcp_nifti_tail` parameters to
refer both to data in HCPpipeline folder structure as well as within
QuNex folder structure. This presented two issues. First, anticipating
addition of other processing pipelines (e.g. fMRIPrep), the `hcp_` prefix
is misleading and confusing. Second, the parameters were used both for
mapping the data as well as for specification of which files to work
with further downstream. These two are in most cases the same, however,
the use of the same parameter for both situations is limiting and could
lead to further challenges downstream.

To resolve the above issues, this commit introduces two new pairs of
parameters. First, to support flexible mapping of data from HCP to QuNex
folder structure `qx_nifti_tail` and `qx_cifti_tail` parameters are now
used. The specific mapping:

    hcp_nifti_tail -> qx_nifti_tail
    hcp_cifti_tail -> qx_nifti_tail

now allows not only copying the correct data but also specifying the
tail to be used within the QuNex environment, in other words the tails
in both environments do not have to match.

Second, to support more flexible specification of the files to work with,
this commit limits the use of the `qx_*_tail` parameters to mapping only
(specifically to `mapHCPData` command) and adds two additional parameters,
`nifti_tail` and `cifti_tail`, which now specify which files to work with
when processing and analysing data in the QuNex environment.

To enable backward compatibility with existing scripts and batch.txt files
that do not specify the new parameters, this commit also introduces new
functionality to processing of parameters, parameter imputation using a new
internal function `imputeParameters`. The function takes a list of tuples
specifying the target and source parameter names and then assigns the value
of the source parameter to the target parameter, if the target parameter is
None.

Using this functionality, if the `qx_nifti_tail` or `qx_cifti_tail` is not
specified, it takes the value of `hcp_nifti_tail` or `hcp_cifti_tail`,
respectively. Similarly, if any of `nifti_tail` or `cifti_tail` is not
specified, it takes the value of `qx_nifti_tail` or `qx_cifti_tail`,
respectively. The parameters are inputed in the order listed above, which
enables the cascade of values:

    hcp_nifti_tail -> qx_nifti_tail -> nifti_tail

In other words, if neither `qx_nifti_tail` nor `nifti_tail` are specifed,
`qx_nifti_tail` will first inherit the value from `hcp_nifti_tail` and then
pass it on to `nifti_tail` as well.

trello-card: https://trello.com/c/CBVBi3WU