fix: make use of parameter values in batchTag2NameKey
batchTag2NameKey did not save parameter values obtained from the batch file
and it did not pass them on to useOrSkipBOLD. This led to an exception in
the case where some bolds were ignored, as useOrSkipBOLD did not have
information on 'hcp_filename' parameter.

The fix now passes the parameters read from the batch file on to
useOrSkipBOLD and it also makes the useOrSkipBOLD robust to a missing
'hcp_filename' parameter.

Trello-card: https://trello.com/c/T0xefats