feat: Add the option to use a custom mask in run_palm
In some cases, when running PALM, the user might want to use a custom mask
that specifies which voxels or grayordinates the analysis is to be
restricted to. This is now enabled by specifying 'mask:<path to mask>'
option. When specified, the mask file will be copied to the working folder
and if needed (for dtseries and dscalar files) deconstructed into
constituent left, right and volume components before use. The temporary
mask file is removed after PALM completes.

trello-card: https://trello.com/c/vPHWeR50