fix: ensure that octave is reading empty lines correctly in general_read_table
If there is an empty line in a file, octave would skip that line when
splitting the file into lines using `textscan`, whereas matlab would
read that line as an empty string. The latter enables correct counting
of lines before data in the input table file. The code now uses
`strsplit` with 'CollapseDelimiters' set to false instead.