feat: add information on timeseries frames extracted
When extracting timeseries, bad frames can be ommitted. To perserve the information
on which frames were extracted, a new field `tframes` was added to the `nimage`
class. This field is used by `img_extract_timeseries`. When extraction is set to
include all (usable) frames from events or timeeries, the field holds a vector
storing for each frame in the extracted timeseries the index of the frame from the
original bold (or conc) timeseries. When extraction is set to summarise each event
by a single value (e.g. mean), the `tframes` vector holds the number of frames
in the source timeseries that were used to compute the summary value.