fix: correct scaling of the assumed regressors
To ensure unit maximum amplitude 1 for the assumed regressors, the values of
the assumed regressors were divided by the maximum positive value after
convolution of the event vector with the assumed HRF. Normalization was
performed independently for the values within each bold run. This worked
correctly for modeling the block design where the assumed regressors reached
the maximum value. It did not work correctly for short events and for designs
with behavioral weights. If the events were short and the regressor did not
reach the maximum possible value, they were scaled too high: e.g. a regressor
with 1s long events would scale to the maximum value of 1, which would be
disproportionate to regressors with events of longer duration. Similarly, the
regressor would be scaled so that the event with the highest weight would scale
to 1. This could lead to problems if:

* Beta values from different regressors with short events of different lengths
  were to be compared with each other,
* the regressor would span multiple bolds within an individual and would
  contain short events that would differ in length between bold runs,
* the regressor would include short events of different lengths in different
  participants or groups of participants,
* the regressor would include behavioral (or other) weights, it would span￼
  more than one bold run, and have different maximum weights in across bold
  runs.

To address this potential problem, the regressor is now scaled by the area
under the positive part of the HRF. This ensures the same scaling for all
regressors, regardless of the length of the events and/or the presence of
behavioral (or other) weights. The scaling ensures that any event with a weight
of 1 and a￼ length sufficient to reach the maximum amplitude is scaled to 1. All
other events are scaled proportionaly.

trello-card: https://trello.com/c/ecynEWQp