fix: set no-conf2hires to be passed as flag
When `hcp_fs_no_conf2hires` QuNex parameter was set to TRUE, the resulting
FreeSurferPipeline call treated `--no-conf2hires` as a parameter and set it to
`--no-conf2hires=TRUE` instead of using it as a flag an including in the call
as `--no-conf2hires`only. The behavior is now changed to a correct one.

trello-card: https://trello.com/c/89BBcgjh