feat: enable selection of assumed regressor normalization
A change in normalization of assumed regressors leads to a change in resulting
beta values even in cases where the change itself does not impact the validity
of results. This is problematic in the case of ongoing studies and previously
processed data. To ensure backwards compatibility we added in this commit an
explicit option to normalize the assumed regressors within each run (the old
behavior) or to normalize the area under the curve of the hrf function and so
ensure universal normalization. The default currently remain within run
normalization.

The new functionality is enabled for each assumed regressor using the event
string specification.

The inline documentation has been updated to reflect the new functionality.

trello-card: https://trello.com/c/ecynEWQp