fix: Remove conflicting code in `depreciated_values{}`
The dictionary `depreciated_values` has two pairs of identical keys. The
first two are the updated latest values. The last two are old values
which were probably reintroduced during a merge conflict after `0.91.11`
(which doesn't have this issue).

This regression bug causes the value of `option['hcp_filename']` to not get
changed into `userdefined` in qunex `0.92.2`.