feat: improve reading of file lists
0123456789012345678901234567890123456789012345678901234567890123456789012345
When reading file lists it is important to check that all the required files
are present for each session. This update extends the checking ability by
adding a parameter that specifies, which elements have to be present. The
function now also returns a list of missing elements for each of the sessions
in the list file.

trello-card: https://trello.com/c/oRQdPOfq