feat: enable computation of predicted and residual BOLD timeseries
After GLM analysis of BOLD data, it is sometimes needed to compute either a
predicted timeseries based on the GLM results, or to compute a residual
timeseries after one or a number of regressors are regressed out of the
original BOLD timeseries. This commit adds the functionality to achieve that.
It provides two new Matlab functions, img_glm_predict and general_glm_predict.
The first is a nimage method, the second is a script that enables running
glm_predict across a set of sessions as they are provided in a list file.
general_glm_predict is added to the list of matlab functions accessible
through gmri/qunex command.

trello-card: https://trello.com/c/oRQdPOfq