fix: correct computation of two-tail conjunction analyses
When running conjunction analyses on two-tail p-values or z-values the
results might be incorrect. The functionality was changed to deal with
twotailed results. Specifically, z-value input is treated unchanged and the
correction for twotailed test is applied at FDR thresholding. For p-values,
the twotailed p-values are first converted to one-tailed p-values before
computation of pooled p-values. The corection for twotailed test is again
performed before FDR thresholding.

In addition, a new parameter, psign, allows specifying the direction of the
changes in the original twotailed test when undirected p-values are provided
as input. For z-values the sign of the z-values is treated as signifying the
direction of differences.

Third, when saving the results, the function now saves both unthresholded
and thresholded p-values and z-values.

The commit adjusts the documentation and the call of the function from
gmri/qunex.

trello-card: https://trello.com/c/oRQdPOfq