fix: correct printout of executed external commands
To prettify the report on which command was run, the call string was
processed before printout with the intention of printing each parameter on
a separate line. The processing falsely removed spaces and introduced other
errors in the command printout. For that reason the reported call printout
was incorrect and the user could not copy-paste it to investigate and debug
when issues with the executed call occured. This commit turns off additional
processing of the call string before it is printed out.

trello-card: https://trello.com/c/oRQdPOfq