fix: strip sequence specific information in session.txt file
When some sequence specific information was written to session.txt file unnecessary
spaces were added to the entries. These spaces were not stripped when the information
was later read in resulting in errors. The fix changes the writing so that there
are no extraneous spaces included, and it also strips any extra whitespace when
reading.