hotfix: remove tfile before running ReApplyFix
Currently the test file for ReApplyFix is not removed and HCP ReApplyFix
is not executed if a test file is present, even if the overwrite flag is
set to True. Similarly as for executeHCPHandReclassification the fix
removes the test file before running, as any call for ReApplyFix is
valid and overwrite is implicitly True.