feat: process json files in bids import
Json sidecar files are now copied to the nii folder along with image
files. Information specified is added from json to sequence information
in session.txt file. Fieldmap sequences are now tagged with fm(N) or
se(N) information and that same information is added to sequences that
the field maps are intended for (as specifed in the json IntendedFor
field.