feat: optimize checking of image type
This commit adds a `general_check_image_file` function that simplifies
checking of whether a file name is a valid image file name and returns
information of the type of image file and whether the file exists.

Filetype information now includes NIfTI and is stored without the '.'
prefix. All the functions that use filetype information are adjusted to
take that into account.