fix: update nifti header
After the latest changes, nifti header was not updated upon save, which
led to issues if the image changed size. For cifti images the preparation
of nifti header is now left to cifti-matlab. There was an issue of
difference in intent_name size (15 vs 16 characters) and the intent_name
is checked to be of the appropriate length