update: optimize processing of file names
The object now stores only the full filename of the image in
`filenamepath` and `filenamepaths` variables. All other versions of
the filename are obtained using the object methods:

img_filename() - only the filename without the path
img_filenamepaths() - the full filename including the path
img_basename() - the name of the file w/o the image type extensions
img_basenamepath() - the name of the file w/o the extensions and with
full path
img_fileext() - the full image-related extension of the file (e.g.,
`.dtseries.nii`)
img_filetype() - the type of the image only (e.g., `dtseries`, `nifti`)

All of the above methods also accept a filenamepath as a variable, if
the information needs to be extracted from an arbitrary path instead of
the actual filenamepath of the image object.