feat: enable running join_fidl on cifti data
join_fidl and join_fidl_folder relied on number of frames reported for bold
files based on nifti header. This was a problem as img library did not support
reading NIfTI 2 headers. This commit builds on previous commit enabling
reading of NIfTI 2 headers, and adds code to (i) get the number of
volumes instead of frames and use it to compute bold length
when joining fidl files.